Adding servo to auto
Fixed bad comment on servo open/close values.

Added claw servo to Auto to see if actuating it to closed position made the hang more reliable.  It did not seem to make a difference.  Leaving it in since we may want to still use claw servo in auto in the future.